Implement visitor pattern to SyntaxTree::Node
This pattern will expand our horizons and allow us to do fun stuff with the nodes through the SyntaxTree::Visitor class.

We've chosen to require the Visitor class in node.rb to make it clear that the file requires it. Otherwise, we'd have to add it to lib/syntax_tree.rb in an order-dependent way, which is undesirable.

Co-authored-by: Kaan Ozkan <kaan.ozkan@shopify.com>
Co-authored-by: Alexandre Terrasa <alexandre.terrasa@shopify.com>
Co-authored-by: Vinicius Stock <vinicius.stock@shopify.com>
Co-authored-by: Emily Giurleo <emily.giurleo@shopify.com>
Co-authored-by: Rafael Franca <rafael.franca@shopify.com>