Add support for a simple config file
As syntax_tree gains wider usage, it'll be very helpful to allow
projects to commit their particular configs within their repo for use by
any tool that uses `stree` instead of having to create wrapper scripts,
rely on rake tasks that come with their own overhead, or manually keep
settings up to date among all contributors.

This set of changes takes inspiration from sorbet to provide users with
a simple config file that's also easy to parse and use within the gem.
It's a text file that has the exact options you'd pass on the command
line, each on their own line. These are parsed and prepended to the
arguments array within `CLI.run`, still allowing for other options to
passed from the command line.

I decided to restrict this only to the command line options and avoid
the source files argument, opting to let other tools pass their own
source file from the command line, which is preferable for tools like
editor integrations that might interact with a single file at a time. If
users want to interact with all of their Ruby files at once, the rake
tasks are perfect for providing larger, static patterns of files. And
since they use `CLI.run` as well, they'll pick up options from a
.streerc file, if present.

I also opted for only supporting a single .streerc file at the project
root. If there's a need for multiple configs or config locations, this
can be easily extended to look up through a directory structure or
accept an option for a specific config file location (or even a
different filename). Those felt out of scope for this initial support.