Fix CLI checking for content
Previously, we were checking if $stdin was a TTY to determine if
there was content to be read. As it turns out, this isn't really a
good indicator, as content could always come later, and some folks
run stree in CI when $stdin is not a TTY and still pass filenames.

Instead, we now check if no filenames were passed, and in that case
we attempt to read from $stdin.