Fix Ruby build on HEAD
This fixes two bugs. The first is that in an `in` clause, you need
to use the `then` keyword or parentheses if the pattern you are
matching is an endless range.

The second is that we are associating the `then` keyword with the
wrong `in` clauses because they come in in reverse order and we're
deleting them from the parent clauses incorrectly.